refactored Claim4 to a more meaningful structure
First of all, Claim 4 *works*.  However, since the output was not very
readable, it would discourage outsiders.  Therefore, we internally apply
the best practice of not connecting activities across concurrent regions
and decision nodes.  We always have an activity before a decision
instead.  See Claim 5 for the refactored version of Claim 4.